Bug Fix
Bug fix, can't connect after change transmission type from tcp to other types

If flow is not equal to "" in transmit except tcp, connection will not be made. This problem occurs when we change the transmission from tcp to another type. Also, if tcp itself is alone and without tls and reality, flow must be empty in it.